feat: migrate to new design system & theme switch (FSD architecture)

+++ apps/taskmaster-server/src/services/task.service.ts
@@ -1,6 +1,6 @@
 import { TaskStatus } from "@prisma/client";
 import { prisma } from "../prisma/client";
-import e from "express";
+import { deleteObjectFromS3 } from "../lib/s3Client";
 
 interface TaskInput {
   boardId: string;
@@ -123,6 +123,12 @@ export const getTask = async (taskId: string) => {
         assignees: {
           include: {
             user: true,
+          },
+        },
+        notes: {
+          include: {
+            author: true,
+            file: true,
           },
         },
       },
@@ -248,6 +254,7 @@ export const deleteTask = async (taskId: string, currentUserId: string) => {
             ownerId: true,
           },
         },
+        files: true,
       },
     });
 
@@ -265,10 +272,22 @@ export const deleteTask = async (taskId: string, currentUserId: string) => {
       );
     }
 
+    for (const file of task.files) {
+      if (file.key) {
+        try {
+          await deleteObjectFromS3(file.key);
+          console.log(`File ${file.key} deleted from S3 successfully`);
+        } catch (error) {
+          console.error("Error deleting file from S3:", error);
+        }
+      }
+    }
+
+    await prisma.taskFile.deleteMany({ where: { taskId } });
     await prisma.taskAssignee.deleteMany({ where: { taskId } });
     await prisma.task.delete({ where: { id: taskId } });
 
-    return { message: "Task deleted successfully" };
+    return { message: "Task and its files deleted successfully" };
   } catch (error: any) {
     console.error("Error deleting task:", error);
     throw new Error(error.message || "Failed to delete task");
@@ -278,7 +297,8 @@ export const deleteTask = async (taskId: string, currentUserId: string) => {
 export const createNote = async (
   taskId: string,
   content: string,
-  userId: string
+  userId: string,
+  fileId?: string
 ) => {
   try {
     const isAssignee = await prisma.taskAssignee.findFirst({
@@ -291,16 +311,19 @@ export const createNote = async (
     if (!isAssignee) {
       throw new Error("User is not an assignee of this task");
     }
+
     const note = await prisma.taskNote.create({
       data: {
         content,
         taskId,
         authorId: userId!,
+        fileId: fileId || undefined,
       },
       include: {
         author: {
           select: { id: true, name: true },
         },
+        file: true,
       },
     });
 
@@ -311,18 +334,8 @@ export const createNote = async (
   }
 };
 
-export const getNotesByTask = async (taskId: string, userId: string) => {
+export const getNotesByTask = async (taskId: string) => {
   try {
-    // const isAssignee = await prisma.taskAssignee.findFirst({
-    //   where: {
-    //     taskId,
-    //     userId,
-    //   },
-    // });
-
-    // if (!isAssignee) {
-    //   throw new Error("User is not an assignee of this task");
-    // }
     const notes = await prisma.taskNote.findMany({
       where: { taskId },
       orderBy: { createdAt: "asc" },
@@ -333,6 +346,7 @@ export const getNotesByTask = async (taskId: string, userId: string) => {
             name: true,
           },
         },
+        file: true,
       },
     });
 
@@ -359,6 +373,12 @@ export const updateNote = async (
   const updatedNote = await prisma.taskNote.update({
     where: { id: noteId },
     data: { content },
+    include: {
+      author: {
+        select: { id: true, name: true },
+      },
+      file: true,
+    },
   });
 
   return updatedNote;
@@ -368,14 +388,19 @@ export const deleteNote = async (noteId: string, userId: string) => {
   try {
     const existingNote = await prisma.taskNote.findUnique({
       where: { id: noteId },
+      include: { file: true },
     });
 
     if (!existingNote || existingNote.authorId !== userId) {
       throw new Error("You can only delete your own notes");
     }
 
-    await prisma.taskNote.delete({ where: { id: noteId } });
+    if (existingNote) await prisma.taskNote.delete({ where: { id: noteId } });
 
+    if (existingNote.file?.key && existingNote.fileId) {
+      await deleteObjectFromS3(existingNote.file.key);
+      await prisma.taskFile.delete({ where: { id: existingNote.fileId! } });
+    }
     return { message: "Note deleted successfully" };
   } catch (error: any) {
     console.error("Error deleting note:", error);
